moved functions to utils, so state can have a state.change(params) function.
This is a move to allow state to manage multiple stacks of slides or views. to allow different jump order depending on mode
